The default example feature is now behind a flag
