+++ core/src/main/kotlin/in/specmatic/core/Flags.kt
@@ -6,6 +6,7 @@ object Flags {
     private const val customResponseName = "CUSTOM_RESPONSE"
     const val negativeTestingFlag = "SPECMATIC_GENERATIVE_TESTS"
     const val maxTestRequestCombinationsFlag = "MAX_TEST_REQUEST_COMBINATIONS"
+    const val schemaExampleDefault = "SCHEMA_EXAMPLE_DEFAULT"
 
     private fun flagValue(flagName: String): String? {
         return System.getenv(flagName) ?: System.getProperty(flagName)
@@ -13,6 +14,10 @@ object Flags {
 
     fun customResponse(): Boolean {
         return flagValue(customResponseName) == "true"
+    }
+
+    fun schemaExampleDefaultEnabled(): Boolean {
+        return BooleanUtils.toBoolean(flagValue(schemaExampleDefault) ?: "true")
     }
 
     fun generativeTestingEnabled(): Boolean {
