+++ core/src/main/kotlin/in/specmatic/core/pattern/Pattern.kt
@@ -1,5 +1,6 @@
 package `in`.specmatic.core.pattern
 
+import `in`.specmatic.core.Flags
 import `in`.specmatic.core.Resolver
 import `in`.specmatic.core.Result
 import `in`.specmatic.core.value.StringValue
@@ -60,4 +61,20 @@ interface Pattern {
     val typeAlias: String?
     val typeName: String
     val pattern: Any
+}
+
+fun matchingExample(example: String?, pattern: Pattern): Value? {
+    if(!Flags.schemaExampleDefaultEnabled())
+        return null
+
+    if(example == null)
+        return example
+
+    val value = pattern.parse(example, Resolver())
+    val exampleMatchResult = pattern.matches(value, Resolver())
+
+    if(exampleMatchResult.isSuccess())
+        return value
+
+    throw ContractException("Example \"$example\" does not match ${pattern.typeName} type")
 }
